3.8 Validaciones terminado

+++ Segundo trimestre/spring_boot_web_mvc_jdbc_ventas-master/src/main/java/org/iesbelen/dao/ComercialDAOImpl.java
@@ -38,7 +38,7 @@ public class ComercialDAOImpl implements ComercialDAO {
                         rs.getString("nombre"),
                         rs.getString("apellido1"),
                         rs.getString("apellido2"),
-                        rs.getFloat("comisión"))
+                        rs.getBigDecimal("comisión"))
 
         );
 
@@ -56,7 +56,7 @@ public class ComercialDAOImpl implements ComercialDAO {
                                 rs.getString("nombre"),
                                 rs.getString("apellido1"),
                                 rs.getString("apellido2"),
-                                rs.getFloat("comisión"))
+                                rs.getBigDecimal("comisión"))
                         , id
                 );
